Update membership timeframe.
Wharf requested that the timeframe of memberships change. They required
the membership to have a minimum length of 5 months and for expiry
still to happen in May.

Therefore, members joining between May and January will expire in May of
the following year. However, those who join between January to May
should expire in the May of the following year.

This can simply be done by setting the year of the membership to the
calandar year, as opposed to using the this_year variable which
accounts for May being the expiry date of the year.

This means a few simple changes to definitions of "current", which is
no longer equal to this year, but is greater than or equal to this year.

Similarly the query for expired should now return members where the is
no membership year greater than or equal to this year.